### 2025-02-20 DorianDepriester/Elasticipy#1 @Kevin-Mattheus-Moerman
Including the Kelvin mapping in this package.
I am speaking here not as a JOSS editor (over at https://github.com/openjournals/joss-reviews/issues/7802#issuecomment-2670838874) but as a fellow researcher in continuum mechanics, have you considered the Kelvin (sometimes referred to a Mandel) mapping of the Elasticity tensor? I think it is a largely overlooked mapping in solid mechanics but is superior to Voigt, which I see you include. It allows one to get the up to 6 "eigenstrains" (eigen deformation modes) and "eigenstiffnesses" from the elasticity tensor. Also it maintains tensor algebra which the Voigt mapping looses (array multiplication).

Some relevant papers:
[10.1111/j.1365-2478.2011.01049.x](https://doi.org/10.1111/j.1365-2478.2011.01049.x)
[10.1093/qjmam/43.1.15](https://doi.org/10.1093/qjmam/43.1.15)
[10.1007/s12665-016-5429-4](https://doi.org/10.1007/s12665-016-5429-4)

This is a nice presentation on it to:
https://geo.mff.cuni.cz/magma/magma-051102.pps
Where it is also made clear that the general mapping rule is the same as Voigt but it features factors of $\sqrt{2}$.

Note the above is not a requirement for the JOSS review. I merely state it as an interesting potential feature of this package.
