### 2025-02-20 #7802 @Kevin-Mattheus-Moerman
As a side note I posted a question here too: https://github.com/DorianDepriester/Elasticipy/issues/1
