### 2025-02-23 #7802 @DorianDepriester
> [@DorianDepriester](https://github.com/DorianDepriester) Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:
>
> * [ ]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
>
> As a side note I posted a question here too: [DorianDepriester/Elasticipy#1](https://github.com/DorianDepriester/Elasticipy/issues/1)

@Kevin-Mattheus-Moerman, I have added DOIs for [Pymatgen] and [MTEX]. Still, as references [Hill, 1952] and [Nye, 1985] are quite old, no DOIs were attached to them.
