### 2025-02-23 DorianDepriester/Elasticipy#1 @DorianDepriester
Hello @Kevin-Mattheus-Moerman ,
Thank you for this suggestion. It would indeed be an interesting feature, I will investigate the Kelvin convention. Although I am not familiar with it, AFAIK, [CraFT](https://lma-software-craft.cnrs.fr/) uses this convention.
