### 2025-02-24 56a6fa0
Fix error when flipping the eigen vectors
