### 2025-02-24 0bf2ca8
Implement Kevlin, eig, eig_stiffnesses and eig_strains
