### 2025-02-25 4cc5a58
Add test for eigenstiffnesses
