### 2025-02-25 9977060
Create generic function to build a tensor from vectors Add from_Kelvin class method
