### 2025-02-25 7de2a06
Add docstring for from_Kelvin
