### 2025-02-25 76e6d8d
In test_eigenstiffnesses, also test if eigenstiffness are (true) eigenvalues (and also test from_Kelvin)
