### 2025-02-26 5072615
Implement eig() (and so) methods for ComplianceTensor
