### 2025-02-26 05f26e1
Add missing @property
