### 2025-02-26 cdbbecc
rename Kelvin to "to_Kelvin" and remove @property decorator
