### 2025-02-26 a4609e8
Move test to test_StiffnessTensor.py
