### 2025-02-26 0db57bf
Add from_Kelvin() classmethods
