### 2025-02-26 3164089
Add test for conversion from/to Kelvin mapping
