### 2025-02-26 06ddd13
Add to_Voigt method()
