### 2025-02-26 8f12a11
Add test for to_Voigt()
