### 2025-02-26 7121faf
Add to_Voigt() method
