### 2025-02-26 98c0b84
Replace deprecated matmul with dot/rotate with mode='cross'
