### 2025-02-26 6a3afb1
[PEP] Use CamelCase for classname
