### 2025-02-26 a8a91a1
Add JOSS badge to README.md
