### 2025-02-26 2575ae0
Add JOSS badge to index.rst
