### 2025-02-27 DorianDepriester/Elasticipy#1 @Kevin-Mattheus-Moerman
Looks great, thanks. You may find it useful to study anisotropy too.
