### 2025-02-28 43f295f
Use logo instead of plain text
