### 2025-02-28 dcce168
Remove duplicate when listing the tensors
