### 2025-02-28 6862726
Move logo folder to doc/source/
