### 2025-02-28 f7b0372
Add favicon
