+++ docs/source/conf.py
@@ -41,6 +41,7 @@ language = 'english'
 html_theme = "sphinx_rtd_theme"
 html_static_path = ['../_static']
 html_logo = "logo/logo_text_whitebg.png"
+html_favicon = 'logo/favicon.png'
 numpydoc_class_members_toctree = False
 autoclass_content = 'both'
 
