### 2025-03-03 0fb5326
Implement constructor (classmethod) for identity tensor
