### 2025-03-04 624c6a4
Implement ddot for SymmetricTensor
