### 2025-03-04 2e07c07
overload mult ('*') with SymmetricTensor
