### 2025-03-04 48103fb
Add test for ddot
