### 2025-03-05 9a5e024
Make rotate_tensor public
