+++ src/Elasticipy/FourthOrderTensor.py
@@ -103,8 +103,23 @@ def _check_definite_positive(mat):
         raise ValueError('The input matrix is not definite positive (eigenvalues: {})'.format(eigen_val))
 
 
-def _rotate_tensor(full_tensor, rotation):
-    rot_mat = rotation_to_matrix(rotation)
+def rotate_tensor(full_tensor, r):
+    """
+    Rotate a (full) fourth-order tensor.
+
+    Parameters
+    ----------
+    full_tensor : numpy.ndarray
+        array of shape (3,3,3,3) or (...,3,3,3,3) containing all the components
+    r : scipy.spatial.Rotation or orix.quaternion.Rotation
+        Rotation, or set of rotations, to apply
+
+    Returns
+    -------
+    numpy.ndarray
+        Rotated tensor. If r is an array, the corresponding axes will be added as first axes in the result array.
+    """
+    rot_mat = rotation_to_matrix(r)
     str_ein = '...im,...jn,...ko,...lp,mnop->...ijkl'
     return np.einsum(str_ein, rot_mat, rot_mat, rot_mat, rot_mat, full_tensor)
 
@@ -230,7 +245,7 @@ class SymmetricTensor:
         if self.orientations is None:
             return m
         else:
-            return _rotate_tensor(m, self.orientations)
+            return rotate_tensor(m, self.orientations)
 
     def flatten(self):
         """
@@ -274,7 +289,7 @@ class SymmetricTensor:
             Rotated tensor
         """
         if _is_single_rotation(rotation):
-            rotated_tensor = _rotate_tensor(self.full_tensor(), rotation)
+            rotated_tensor = rotate_tensor(self.full_tensor(), rotation)
             rotated_matrix = self._full_to_matrix(rotated_tensor)
             return self.__class__(rotated_matrix)
         else:
