### 2025-03-06 f662364
Allow to force symmetry of tensor
