### 2025-03-06 9f0817a
Add SelfConsistent.py to cvs
