+++ Examples/SelfConsistent.py
@@ -1,0 +1,102 @@
+import numpy as np
+from Elasticipy.FourthOrderTensor import StiffnessTensor, SymmetricTensor
+from scipy.integrate import trapezoid
+from Elasticipy.FourthOrderTensor import rotate_tensor
+from scipy.spatial.transform import Rotation
+
+def ddot(a, b):
+    if isinstance(a, SymmetricTensor):
+        a_full = a.full_tensor()
+    else:
+        a_full = a
+    if isinstance(b, SymmetricTensor):
+        b_full = b.full_tensor()
+    else:
+        b_full = b
+    return np.einsum('...ijmn,...mnkl->...ijkl', a_full, b_full)
+
+def invert_4th_order_tensor(T):
+    shape = T.shape
+    *a,_,_,_,_ = shape
+    T_mat = T.reshape(tuple(a) + (9, 9))
+    T_inv_mat = np.linalg.pinv(T_mat)
+    return T_inv_mat.reshape(shape)
+
+def Kroner_matrix(L, theta, phi, a1=1., a2=1., a3=1.):
+    theta = np.atleast_2d(theta)
+    phi = np.atleast_2d(phi)
+    s1 = np.sin(theta)*np.cos(phi) / a1
+    s2 = np.sin(theta)*np.sin(phi) / a2
+    s3 = np.cos(theta) / a3
+    s = [s1, s2, s3]
+    K = np.array([[L[0,0]*s1**2+L[5,5]*s2**2+L[4,4]*s3**2   ,(L[0,1]+L[5,5])*s1*s2                  ,(L[0,2]+L[4,4])*s1*s3],
+                  [(L[0,1]+L[5,5])*s1*s2                    ,L[5,5]*s1**2+L[1,1]*s2**2+L[3,3]*s3**2 ,(L[1,2]+L[3,3])*s2*s3],
+                  [(L[0,2]+L[4,4])*s1*s3                    ,(L[1,2]+L[3,3])*s2*s3                  ,L[4,4]*s1**2+L[3,3]*s2**2+L[2,2]*s3**2]])
+    return np.einsum('ikmn,jmn,lmn->ijklmn', np.linalg.inv(K.T).T, s, s)
+
+def Morris_tensor(L):
+    gamma = Kroner_matrix(L, theta, phi)
+    a = trapezoid(gamma*np.sin(theta), theta[0], axis=-1)
+    return trapezoid(a, phi[:,0], axis=-1)/(4*np.pi)
+
+def localization_tensor(C_macro, C_incl, orientation):
+    E = Morris_tensor(C_macro.matrix)
+    E_local = rotate_tensor(E, orientation.inv())
+    C_macro_local = rotate_tensor(C_macro.full_tensor(), orientation.inv())
+    I = StiffnessTensor.identity(return_full_tensor=True)
+    Ainv = ddot(E_local, C_incl.full_tensor() - C_macro_local) + I
+    A = invert_4th_order_tensor(Ainv)
+    return rotate_tensor(A, orientation)
+
+def global_spherical_grid(n_theta=50, n_phi=100):
+    global phi, theta
+    theta = np.linspace(0, np.pi, n_theta)
+    phi = np.linspace(0, 2 * np.pi, n_phi)
+    phi, theta = np.meshgrid(phi, theta, indexing='ij')
+
+def Kroner_Eshelby(C, orientations, method='stress',  max_iter=50, atol=1e-3, rtol=1e-4, display=False):
+    C_rotated = C * orientations
+    C_macro = C.Hill_average()
+    global_spherical_grid()
+    keep_on = True
+    iter = 0
+    while keep_on:
+        A = np.zeros((m,3,3,3,3))
+        for i in range(m):
[…]
