### 2025-03-06 33e53a6
Make Kroner_Eshelby suitable for any symmetry
