### 2025-03-06 7c3dcd3
Put I as a global variable
