### 2025-03-06 41e99f1
%Use eigenstiffnesses as a stopping criterion
