### 2025-03-06 c6a54c5
Stop if the max number of iter is reached
