### 2025-03-06 108e7af
Make rotate_tensor compatible with arrays of shape (...,3,3,3,3)
