### 2025-03-07 54e0588
Rotate A (or B) before averaging
