+++ Examples/SelfConsistent.py
@@ -27,18 +27,16 @@ def invert_4th_order_tensor(T):
     T_inv_mat = np.linalg.pinv(T_mat)
     return T_inv_mat.reshape(shape)
 
-def Kroner_tensor(L, theta, phi, a1=1., a2=1., a3=1.):
-    theta = np.atleast_2d(theta)
-    phi = np.atleast_2d(phi)
+def Kroner_tensor(C_macro_local, a1=1., a2=1., a3=1.):
     s1 = np.sin(theta)*np.cos(phi) / a1
     s2 = np.sin(theta)*np.sin(phi) / a2
     s3 = np.cos(theta) / a3
     s = [s1, s2, s3]
-    D = np.einsum('kijl,kpq,lpq->ijpq', L, s, s)
+    D = np.einsum('kijl,kpq,lpq->ijpq', C_macro_local, s, s)
     return np.einsum('ikmn,jmn,lmn->ijklmn', np.linalg.inv(D.T).T, s, s)
 
-def Morris_tensor(L):
-    gamma = Kroner_tensor(L, theta, phi)
+def Morris_tensor(C_macro_local):
+    gamma = Kroner_tensor(C_macro_local)
     a = trapezoid(gamma*np.sin(theta), theta[0], axis=-1)
     return trapezoid(a, phi[:,0], axis=-1)/(4*np.pi)
 
@@ -53,29 +51,29 @@ def global_spherical_grid(n_theta=50, n_phi=100):
     phi = np.linspace(0, 2 * np.pi, n_phi)
     phi, theta = np.meshgrid(phi, theta, indexing='ij')
 
-def Kroner_Eshelby(C, orientations, method='stress',  max_iter=50, atol=1e-3, rtol=1e-4, display=False):
-    C_rotated = C * orientations
+def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, display=False):
+    C_rotated = C * g
     C_macro = C_rotated.Hill_average()
     eigen_stiff = C_macro.eig_stiffnesses
     global_spherical_grid()
     keep_on = True
-    iter = 0
-    reason = 'Maximum number of iterations is reached'
-    m = len(orientations)
+    k = 0
+    message = 'Maximum number of iterations is reached'
+    m = len(g)
     while keep_on:
         eigen_stiff_old = eigen_stiff
         A_local = np.zeros((m,3,3,3,3))
-        C_macro_local = C_macro * orientations.inv()
+        C_macro_local = C_macro * g.inv()
         for i in range(m):
             A_local[i] = localization_tensor(C_macro_local[i].full_tensor(), C)
         LiAi_local = ddot(C, A_local)
         if method == 'stress':
-            LiAi = rotate_tensor(LiAi_local, orientations)
+            LiAi = rotate_tensor(LiAi_local, g)
             C_matrix = np.mean(LiAi, axis=0)
             C_macro = StiffnessTensor(C_matrix, force_symmetry=True)
         elif method == 'strain':
             B_local = ddot(LiAi_local, C_macro_local.inv().full_tensor())
-            Bi = rotate_tensor(B_local, orientations)
+            Bi = rotate_tensor(B_local, g)
             LiBi_mean = np.mean(ddot(C_rotated.inv(), Bi),axis=0)
             S_macro = ComplianceTensor(LiBi_mean, force_symmetry=True)
             C_macro = S_macro.inv()
@@ -86,27 +84,27 @@ def Kroner_Eshelby(C, orientations, method='stress',  max_iter=50, atol=1e-3, rt
         abs_change = np.abs(eigen_stiff - eigen_stiff_old)
         rel_change = np.max(abs_change / eigen_stiff_old)
[…]
