### 2025-03-07 c095ee8
only set (no reset) A_local
