### 2025-03-07 a4fc71e
Rename Kroner_tensor > gamma
