+++ Examples/SelfConsistent.py
@@ -4,10 +4,8 @@ from scipy.integrate import trapezoid
 from Elasticipy.FourthOrderTensor import rotate_tensor
 from scipy.spatial.transform import Rotation
 
-from tests.test_StiffnessTensor import rotations
-
 I = StiffnessTensor.identity(return_full_tensor=True)
-
+#I = np.einsum('ij,kl->ijkl', np.eye(3), np.eye(3))
 
 def ddot(a, b):
     if isinstance(a, SymmetricTensor):
@@ -18,7 +16,7 @@ def ddot(a, b):
         b_full = b.full_tensor()
     else:
         b_full = b
-    return np.einsum('...ijmn,...nmkl->...ijkl', a_full, b_full)
+    return np.einsum('...ijmn,...mnkl->...ijkl', a_full, b_full)
 
 def invert_4th_order_tensor(T):
     shape = T.shape
@@ -27,7 +25,7 @@ def invert_4th_order_tensor(T):
     T_inv_mat = np.linalg.pinv(T_mat)
     return T_inv_mat.reshape(shape)
 
-def Kroner_tensor(C_macro_local, a1=1., a2=1., a3=1.):
+def gamma(C_macro_local, a1=1., a2=1., a3=1.):
     s1 = np.sin(theta)*np.cos(phi) / a1
     s2 = np.sin(theta)*np.sin(phi) / a2
     s3 = np.cos(theta) / a3
@@ -36,8 +34,8 @@ def Kroner_tensor(C_macro_local, a1=1., a2=1., a3=1.):
     return np.einsum('ikmn,jmn,lmn->ijklmn', np.linalg.inv(D.T).T, s, s)
 
 def Morris_tensor(C_macro_local):
-    gamma = Kroner_tensor(C_macro_local)
-    a = trapezoid(gamma*np.sin(theta), theta[0], axis=-1)
+    g = gamma(C_macro_local)
+    a = trapezoid(g*np.sin(theta), theta[0], axis=-1)
     return trapezoid(a, phi[:,0], axis=-1)/(4*np.pi)
 
 def localization_tensor(C_macro_local, C_incl):
@@ -71,15 +69,18 @@ def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, dis
             LiAi = rotate_tensor(LiAi_local, g)
             C_matrix = np.mean(LiAi, axis=0)
             C_macro = StiffnessTensor(C_matrix, force_symmetry=True)
+            AB = rotate_tensor(A_local, g)
         elif method == 'strain':
             B_local = ddot(LiAi_local, C_macro_local.inv().full_tensor())
             Bi = rotate_tensor(B_local, g)
             LiBi_mean = np.mean(ddot(C_rotated.inv(), Bi),axis=0)
             S_macro = ComplianceTensor(LiBi_mean, force_symmetry=True)
             C_macro = S_macro.inv()
+            AB = rotate_tensor(B_local, g)
         else:
             raise ValueError('Only "strain" and "stress" are valid method names')
 
+        # Stopping criteria
         eigen_stiff = C_macro.eig_stiffnesses
         abs_change = np.abs(eigen_stiff - eigen_stiff_old)
         rel_change = np.max(abs_change / eigen_stiff_old)
@@ -94,10 +95,6 @@ def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, dis
         if k == max_iter:
             keep_on = False
         if display:
-            if method == 'stress':
-                AB = rotate_tensor(A_local, g)
-            else:
-                AB = rotate_tensor(B_local, g)
[…]
