### 2025-03-07 d1473a9
Fix wrong definition for identity tensor
