### 2025-03-07 19b8e13
Simplify syntax
