### 2025-03-07 2dfc99d
Put global statement at the beginning
