+++ Examples/SelfConsistent.py
@@ -4,8 +4,8 @@ from scipy.integrate import trapezoid
 from Elasticipy.FourthOrderTensor import rotate_tensor
 from scipy.spatial.transform import Rotation
 
-#I = StiffnessTensor.identity(return_full_tensor=True)
 I = np.einsum('ik,jl->ijkl', np.eye(3), np.eye(3))
+global phi, theta
 
 def ddot(a, b):
     if isinstance(a, SymmetricTensor):
@@ -16,7 +16,7 @@ def ddot(a, b):
         b_full = b.full_tensor()
     else:
         b_full = b
-    return np.einsum('ijmn,...mnkl->...ijkl', a_full, b_full)
+    return np.einsum('...ijmn,...mnkl->...ijkl', a_full, b_full)
 
 def invert_4th_order_tensor(T):
     shape = T.shape
@@ -25,7 +25,7 @@ def invert_4th_order_tensor(T):
     T_inv_mat = np.linalg.inv(T_mat)
     return T_inv_mat.reshape(shape)
 
-def gamma(C_macro_local, a1=10., a2=1., a3=1.):
+def gamma(C_macro_local, a1=1, a2=1, a3=1):
     s1 = np.sin(theta)*np.cos(phi) / a1
     s2 = np.sin(theta)*np.sin(phi) / a2
     s3 = np.cos(theta) / a3
@@ -65,18 +65,17 @@ def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, dis
         for i in range(m):
             A_local[i] = localization_tensor(C_macro_local[i].full_tensor(), C)
         LiAi_local = ddot(C, A_local)
+        LiAi = rotate_tensor(LiAi_local, g)
         if method == 'stress':
-            LiAi = rotate_tensor(LiAi_local, g)
             C_matrix = np.mean(LiAi, axis=0)
             C_macro = StiffnessTensor(C_matrix, force_symmetry=True)
             AB = rotate_tensor(A_local, g)
         elif method == 'strain':
-            B_local = ddot(LiAi_local, C_macro_local.inv().full_tensor())
-            Bi = rotate_tensor(B_local, g)
-            LiBi_mean = np.mean(ddot(C_rotated.inv(), Bi),axis=0)
-            S_macro = ComplianceTensor(LiBi_mean, force_symmetry=True)
+            B = ddot(LiAi, C_macro.inv())
+            LiinvBi = ddot(C_rotated.inv(), B)
+            S_macro = ComplianceTensor(np.mean(LiinvBi,axis=0), force_symmetry=True)
             C_macro = S_macro.inv()
-            AB = rotate_tensor(B_local, g)
+            AB = B
         else:
             raise ValueError('Only "strain" and "stress" are valid method names')
 
@@ -95,7 +94,7 @@ def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, dis
         if k == max_iter:
             keep_on = False
         if display:
-            err = np.max(np.mean(AB, axis=0) - I)
+            err = np.max(np.abs(np.mean(AB, axis=0) - I))
             print('Iter #{}: abs. change={:0.5f}; rel. change={:0.5f}; error={:0.5f}'.format(k, max_abs_change, rel_change,err))
     return C_macro, message
 
@@ -103,5 +102,5 @@ def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, dis
 Cstrip = StiffnessTensor.transverse_isotropic(Ex= 10.2, Ez=146.8, nu_zx=0.274, nu_yx=0.355, Gxz=7)
[…]
