### 2025-03-07 #7802 @srmnitc
👋🏽 @fdmatoz @ngoiz @hakonanes @ThFriedrich would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
