### 2025-03-09 ca9dd53
Set Kelvin as the default mapping convention
