### 2025-03-09 445c7c4
Implement inverse method for genereic SymmetricTensor
