### 2025-03-10 62c51a1
Remove obsolete option check_definite_positive
