### 2025-03-10 f1c5f6b
Move symmetry option to StiffnessTensor
