### 2025-03-10 da4371c
Allow loading matrices of shapes (...,6,6) and (3,3,3,3)
