### 2025-03-10 61515d2
Move Stiffness/Compliance related functions to proper class
