+++ src/Elasticipy/FourthOrderTensor.py
@@ -3,7 +3,7 @@ import re
 
 from Elasticipy.SecondOrderTensor import SymmetricSecondOrderTensor, rotation_to_matrix, is_orix_rotation, \
     SecondOrderTensor, ALPHABET
-from Elasticipy.SecondOrderTensor import _orientation_shape, _is_single_rotation
+from Elasticipy.SecondOrderTensor import _is_single_rotation
 from Elasticipy.StressStrainTensors import StrainTensor, StressTensor
 from Elasticipy.SphericalFunction import SphericalFunction, HyperSphericalFunction
 from scipy.spatial.transform import Rotation
@@ -131,9 +131,6 @@ class FourthOrderTensor:
     ----------
     matrix : np.ndarray
         (6,6) matrix gathering all the components of the tensor, using the Voigt notation.
-    symmetry : str
-        Symmetry of the tensor
-
     """
     tensor_name = 'Symmetric'
     voigt_map = _voigt_to_kelvin_matrix
@@ -141,7 +138,7 @@ class FourthOrderTensor:
     C46_C56_factor = 1.0
     component_prefix = 'C'
 
-    def __init__(self, M, phase_name=None, orientations=None):
+    def __init__(self, M, phase_name=None):
         """
         Construct of stiffness tensor from a (6,6) matrix.
 
@@ -154,8 +151,6 @@ class FourthOrderTensor:
             (3,3,3,3).
         phase_name : str, default None
             Name to display
-        symmetry : str, default Triclinic
-            Name of the crystal's symmetry
         """
         M = np.asarray(M)
         if M.shape[-2:] == (6, 6):
@@ -167,7 +162,6 @@ class FourthOrderTensor:
 
         self.matrix = matrix
         self.phase_name = phase_name
-        self.orientations = orientations
 
         for i in range(0, 6):
             for j in range(0, 6):
@@ -183,25 +177,8 @@ class FourthOrderTensor:
             heading = '{} tensor (in Voigt notation):\n'.format(self.tensor_name)
         else:
             heading = '{} tensor (in Voigt notation) for {}:\n'.format(self.tensor_name, self.phase_name)
-        print_symmetry = '\nSymmetry: {}'.format(self.symmetry)
-        msg = heading + self.matrix.__str__() + print_symmetry
-        if self.orientations is not None:
-            shape = self.shape
-            if len(shape) == 1:
-                msg = msg + '\n{} orientations'.format(shape[0])
-            else:
-                msg = msg + '\n{} orientations'.format(shape)
+        msg = heading + self.matrix.__str__()
         return msg
-
-    def __len__(self):
-        o = self.orientations
-        if o is None:
-            return 1
-        else:
-            if is_orix_rotation(o):
-                return o.shape[0]
-            else:
-                return len(o)
 
     @property
     def shape(self):
@@ -212,11 +189,8 @@ class FourthOrderTensor:
         tuple
             Shape of the tensor array
         """
-        o = self.orientations
-        if o is None:
-            return None
-        else:
-            return _orientation_shape(o)
+        *shape, _, _ = self.matrix.shape
+        return shape
 
     def full_tensor(self):
         """
@@ -230,11 +204,8 @@ class FourthOrderTensor:
         i, j, k, ell = np.indices((3, 3, 3, 3))
         ij = voigt_indices(i, j)
         kl = voigt_indices(k, ell)
-        m = self.matrix[ij, kl] / self.voigt_map[ij, kl]
-        if self.orientations is None:
-            return m
-        else:
-            return rotate_tensor(m, self.orientations)
+        m = self.matrix[..., ij, kl] / self.voigt_map[ij, kl]
+        return m
 
     def flatten(self):
         """
[…]
