### 2025-03-10 0bd232a
Remove any mention of orientation (no longer used)
