### 2025-03-10 5a8c65a
Fix getitem error
