### 2025-03-10 9ab831d
Fix inconsistent ways of flattening
