### 2025-03-10 f686f45
Remove test for len(ght), which is no longer used
