### 2025-03-10 5e40d69
Fix wrong axes when transposing
