### 2025-03-10 66b120f
Inherit properties in flatten
