### 2025-03-10 44077a7
Inherit properties in rotate
