### 2025-03-10 22e40ce
Inherit properties in rotate
