### 2025-03-10 01045e0
Ensure that the anisotropy factor is computed on the single tensor only
