### 2025-03-10 e0f5e1b
mean now returns a FourthOrderTensor
