### 2025-03-10 518256e
Ensure that mean() is performed on the ndim first axes only
