### 2025-03-10 297b229
Voigt and Reuss now use the mean method
