### 2025-03-10 ee32a3c
Fix undefined variable c
