### 2025-03-10 69e8009
As mean now returns a tensor, use .matri for comparison
