### 2025-03-10 809c3a9
Implement ddot for SecondOrderTensor
