### 2025-03-10 80a7670
Remove former matmul method (replaced by ddot(..., mode='cross')
