### 2025-03-10 1486594
Remove former matmul method (replaced by ddot(..., mode='cross')
