### 2025-03-10 0409f59
Remove error, which is no loger thrown thank to broadcasting
