### 2025-03-10 bfffc5f
Implement zeros constructor
