### 2025-03-10 10562e4
Add docstring for inv()
