### 2025-03-10 7679e40
Remove formed options for SymmetricFourthOrderTensor
