### 2025-03-10 8966134
Implement different definition of identity tensor for SymmetricFourthOrderTensor
