### 2025-03-10 4545ce1
Fix signature of inputs and return
