### 2025-03-10 8c25c27
Fix wrong missing constructor call
