### 2025-03-10 6c67687
Remove 'symmetric' name when printing non-symmetric tensors
