### 2025-03-10 ec8e006
Fix wrong transpose when converting bacvk and forth to full tensor
