### 2025-03-10 a50b100
Use only 'symmetric' definition for identity tensor
