### 2025-03-10 a9e5f07
Allow int as argument for zeros
