### 2025-03-11 ac140ec
mapping is now passed to the constructor
