### 2025-03-11 34029be
C1:C2 now returns a 4th-order tensor
