+++ Examples/Multiple_phases.py
@@ -1,4 +1,4 @@
-from Elasticipy.FourthOrderTensor import StiffnessTensor
+from Elasticipy.Elasticity import StiffnessTensor
 import numpy as np
 from matplotlib import pyplot as plt
 
