+++ Examples/Elasticipy_vs_pymatgen.py
@@ -1,7 +1,7 @@
 from docutils.utils.math.latex2mathml import layout_styles
 
 from Elasticipy.StressStrainTensors import StrainTensor
-from Elasticipy.FourthOrderTensor import StiffnessTensor
+from Elasticipy.Elasticity import StiffnessTensor
 from matplotlib import pyplot as plt
 import matplotlib as mpl
 mpl.use('Qt5Agg')   # Ensure interactive plot
