+++ Examples/SelfConsistent.py
@@ -1,31 +1,11 @@
 import numpy as np
-from Elasticipy.FourthOrderTensor import StiffnessTensor, SymmetricFourthOrderTensor, ComplianceTensor, \
-    FourthOrderTensor
+from Elasticipy.FourthOrderTensor import StiffnessTensor, SymmetricFourthOrderTensor, ComplianceTensor, FourthOrderTensor
 from scipy.integrate import trapezoid
-from Elasticipy.FourthOrderTensor import rotate_tensor
 from scipy.spatial.transform import Rotation
 
 I = FourthOrderTensor.identity()
 #I = np.einsum('ik,jl->ijkl', np.eye(3), np.eye(3))
 global phi, theta
-
-def ddot(a, b):
-    if isinstance(a, SymmetricFourthOrderTensor):
-        a_full = a.full_tensor()
-    else:
-        a_full = a
-    if isinstance(b, SymmetricFourthOrderTensor):
-        b_full = b.full_tensor()
-    else:
-        b_full = b
-    return np.einsum('...ijmn,...mnkl->...ijkl', a_full, b_full)
-
-def invert_4th_order_tensor(T):
-    shape = T.shape
-    *a,_,_,_,_ = shape
-    T_mat = T.reshape(tuple(a) + (9, 9))
-    T_inv_mat = np.linalg.inv(T_mat)
-    return T_inv_mat.reshape(shape)
 
 def gamma(C_macro_local, a1=1, a2=1, a3=1):
     s1 = np.sin(theta)*np.cos(phi) / a1
@@ -44,7 +24,8 @@ def Morris_tensor(C_macro_local):
 
 def localization_tensor(C_macro_local, C_incl):
     E = Morris_tensor(C_macro_local)
-    Ainv = E.ddot(C_incl - C_macro_local) + I
+    delta = FourthOrderTensor(C_incl.full_tensor() - C_macro_local)
+    Ainv = E.ddot(delta) + I
     return Ainv.inv()
 
 def global_spherical_grid(n_theta=50, n_phi=100):
@@ -55,29 +36,30 @@ def global_spherical_grid(n_theta=50, n_phi=100):
 
 def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, display=False):
     C_rotated = C * g
-    C_macro = C_rotated.Hill_average()
+    C_macro = StiffnessTensor.isotropic(E=100, nu=0.3)
     eigen_stiff = C_macro.eig_stiffnesses
     global_spherical_grid()
     keep_on = True
     k = 0
     message = 'Maximum number of iterations is reached'
     m = len(g)
-    A_local = np.zeros((m, 3, 3, 3, 3))
+    A_local = FourthOrderTensor.zeros(m)
     while keep_on:
         eigen_stiff_old = eigen_stiff
         C_macro_local = C_macro * g.inv()
         for i in range(m):
             A_local[i] = localization_tensor(C_macro_local[i].full_tensor(), C)
-        LiAi_local = ddot(C, A_local)
-        LiAi = rotate_tensor(LiAi_local, g)
+        A = A_local * g
+        CiAi = C_rotated.ddot(A)
         if method == 'stress':
-            C_matrix = np.mean(LiAi, axis=0)
-            C_macro = StiffnessTensor(C_matrix, force_symmetry=True)
-            AB = rotate_tensor(A_local, g)
+            LiAi_mean = CiAi.mean().full_tensor()
+            C_macro = StiffnessTensor(LiAi_mean, force_symmetry=True)
+            AB = A
         elif method == 'strain':
-            B = ddot(LiAi, C_macro.inv())
[…]
