### 2025-03-11 a0e1489
Remove duplicate StiffnessTensor
