+++ src/Elasticipy/FourthOrderTensor.py
@@ -501,1192 +501,3 @@ class SymmetricFourthOrderTensor(FourthOrderTensor):
             self.matrix = 0.5*(self.matrix + self.matrix.swapaxes(-1,-2))
         elif check_symmetry and not np.all(np.isclose(self.matrix, self.matrix.swapaxes(-1,-2))):
             raise ValueError('The input matrix must be symmetric')
-
-class StiffnessTensor(SymmetricFourthOrderTensor):
-    """
-    Class for manipulating fourth-order stiffness tensors.
-    """
-    tensor_name = 'Stiffness'
-    C11_C12_factor = 0.5
-    C46_C56_factor = 1.0
-    component_prefix = 'C'
-
-    def __init__(self, M, symmetry='Triclinic', check_positive_definite=True, phase_name= None, mapping='Voigt', **kwargs):
-        """
-        Construct of stiffness tensor from a (6,6) matrix.
-
-        The input matrix must be symmetric, otherwise an error is thrown (except if check_symmetry==False, see below)
-
-        Parameters
-        ----------
-        M : np.ndarray
-            (6,6) matrix corresponding to the stiffness tensor, written using the Voigt notation, or array of shape
-            (3,3,3,3).
-        phase_name : str, default None
-            Name to display
-        symmetry : str, default Triclinic
-            Name of the crystal's symmetry
-        check_symmetry : bool, optional
-            Whether to check or not that the input matrix is symmetric.
-        force_symmetry : bool, optional
-            If true, the major symmetry of the tensor is forces
-        """
-        super().__init__(M, mapping=mapping, **kwargs)
-        if check_positive_definite:
-            _check_definite_positive(self.matrix)
-        self.symmetry = symmetry
-        self.phase_name = phase_name
-
-    def __mul__(self, other):
-        if isinstance(other, StrainTensor):
-            new_tensor = super().__mul__(other)
-            return StressTensor(new_tensor.matrix)
-        elif isinstance(other, StressTensor):
-            raise ValueError('You cannot multiply a stiffness tensor with a Stress tensor.')
-        else:
-            return super().__mul__(other)
-
-    def inv(self):
-        """
-        Compute the reciprocal compliance tensor
-
-        Returns
-        -------
-        ComplianceTensor
-            Reciprocal tensor
-        """
-        C = np.linalg.inv(self.matrix)
-        return ComplianceTensor(C, symmetry=self.symmetry, phase_name=self.phase_name)
-
-    @classmethod
-    def from_txt_file(cls, filename):
-        """
-        Load the tensor from a text file.
-
-        The two first lines can have data about phase name and symmetry, but this is not mandatory.
-
-        Parameters
-        ----------
-        filename : str
-            Filename to load the tensor from.
-
-        Returns
-        -------
-        SymmetricFourthOrderTensor
-            The reconstructed tensor read from the file.
-
-        See Also
-        --------
-        save_to_txt : create a tensor from text file
-
-        """
-        with open(filename, 'r') as f:
-            lines = f.readlines()
-
-        # Initialize defaults
-        phase_name = None
-        symmetry = 'Triclinic'
-        matrix_start_index = 0
-
-        # Parse phase name if available
-        if lines and lines[0].startswith("Phase Name:"):
-            phase_name = lines[0].split(": ", 1)[1].strip()
-            matrix_start_index += 1
-
-        # Parse symmetry if available
-        if len(lines) > matrix_start_index and lines[matrix_start_index].startswith("Symmetry:"):
-            symmetry = lines[matrix_start_index].split(": ", 1)[1].strip()
-            matrix_start_index += 1
-
-        # Parse matrix
[…]
