+++ docs/source/Tutorials/Tutorial_wave-velocities.rst
@@ -11,7 +11,7 @@ We thus start by defining the stiffness tensor for forsterite:
 
 .. doctest::
 
-    >>> from Elasticipy.FourthOrderTensor import StiffnessTensor
+    >>> from Elasticipy.tensors.elasticity import StiffnessTensor
     >>> C = StiffnessTensor.orthorhombic(phase_name='forsterite',
     ...                                  C11=320, C12=68.2, C13=71.6, C22=196.5, C23=76.8,
     ...                                  C33=233.5, C44=64, C55=77, C66=78.7)
