+++ docs/source/Tutorials/Tutorial_MultiplePhases.rst
@@ -42,7 +42,7 @@ Let's start by creating the two stiffness tensors in Python:
 
 .. doctest::
 
-    >>> from Elasticipy.FourthOrderTensor import StiffnessTensor
+    >>> from Elasticipy.tensors.elasticity import StiffnessTensor
     >>> C_austenite = StiffnessTensor.cubic(C11=204, C12=137, C44=126)
     >>> C_ferrite = StiffnessTensor.cubic(C11=242, C12=146, C44=116)
 
