+++ docs/source/Elasticipy.tensor.stress_strain.rst
@@ -1,7 +1,0 @@
-Elasticipy.StressStrainTensors module
-=====================================
-
-.. automodule:: Elasticipy.StressStrainTensors
-   :members:
-   :undoc-members:
-   :show-inheritance:
