### 2025-03-11 a8acb79
Also print symmetry and phase name, if provided
