### 2025-03-11 dabb1c8
Fix doctest ith new formating of __repr__
