+++ docs/source/Tutorials/Tutorial_ReadWriteFiles.rst
@@ -32,6 +32,7 @@ Using the example above:
      [  0.   0.   0.  77.   0.   0.]
      [  0.   0.   0.   0.  77.   0.]
      [  0.   0.   0.   0.   0.  77.]]
+    Phase: Cu
     Symmetry: cubic
 
 
