+++ docs/source/Tutorials/Tutorial_wave-velocities.rst
@@ -23,6 +23,7 @@ We thus start by defining the stiffness tensor for forsterite:
      [  0.    0.    0.   64.    0.    0. ]
      [  0.    0.    0.    0.   77.    0. ]
      [  0.    0.    0.    0.    0.   78.7]]
+    Phase: forsterite
     Symmetry: orthorhombic
 
 And define the mass density:
