### 2025-03-11 51354af
Fix wrong formatting of doctests
