+++ src/Elasticipy/tensors/second_order.py
@@ -552,7 +552,7 @@ class SecondOrderTensor:
 
         Examples
         --------
-        >>> from Elasticipy.SecondOrderTensor import SecondOrderTensor
+        >>> from Elasticipy.tensors.second_order import SecondOrderTensor
         >>> A=SecondOrderTensor.rand(10)
         >>> B=SecondOrderTensor.rand(10)
         >>> AB_pair = A.dot(B)
@@ -1078,7 +1078,7 @@ class SecondOrderTensor:
         --------
         Generate a single random tensor:
 
-        >>> from Elasticipy.SecondOrderTensor import SecondOrderTensor as tensor
+        >>> from Elasticipy.tensors.second_order import SecondOrderTensor as tensor
         >>> tensor.rand(seed=123)
         Second-order tensor
         [[0.68235186 0.05382102 0.22035987]
@@ -1403,7 +1403,7 @@ class SymmetricSecondOrderTensor(SecondOrderTensor):
         --------
         We can create a symmetric tensor by privoding the full matrix, as long it is symmetric:
 
-        >>> from Elasticipy.SecondOrderTensor import SymmetricSecondOrderTensor
+        >>> from Elasticipy.tensors.second_order import SymmetricSecondOrderTensor
         >>> a = SymmetricSecondOrderTensor([[11, 12, 13],[12, 22, 23],[13, 23, 33]])
         >>> print(a)
         Symmetric second-order tensor
@@ -1462,7 +1462,7 @@ class SymmetricSecondOrderTensor(SecondOrderTensor):
 
         Examples
         --------
-        >>> from Elasticipy.SecondOrderTensor import SymmetricSecondOrderTensor
+        >>> from Elasticipy.tensors.second_order import SymmetricSecondOrderTensor
         >>> SymmetricSecondOrderTensor.from_Voigt([11, 22, 33, 23, 13, 12])
         Symmetric second-order tensor
         [[11. 12. 13.]
@@ -1573,7 +1573,7 @@ class SkewSymmetricSecondOrderTensor(SecondOrderTensor):
         --------
         One can construct a skew-symmetric tensor by providing the full skew-symmetric matrix:
 
-        >>> from Elasticipy.SecondOrderTensor import SkewSymmetricSecondOrderTensor
+        >>> from Elasticipy.tensors.second_order import SkewSymmetricSecondOrderTensor
         >>> a = SkewSymmetricSecondOrderTensor([[0, 12, 13],[-12, 0, 23],[-13, -23, 0]])
         >>> print(a)
         Skew-symmetric second-order tensor
