### 2025-03-11 4eb5b32
Ensure that Young_modulus, shear_modulus, linear_compressibility and Poisson_ratio only work for single tensors
