### 2025-03-11 e78e548
Also force wave_velocity to work on single tensors
