### 2025-03-11 3356eb7
AlTest if error is thrown with wave_velocity on non single tensor
