### 2025-03-11 9754736
_mult_ is only permitted if broadcasting applies
