### 2025-03-11 84f5fb4
add test for __mult__
