### 2025-03-11 42df16e
Implement comparison with float/int
