### 2025-03-11 6486ed3
Add test for setting and checking values in array
