+++ tests/test_FourthOrderTensors.py
@@ -40,6 +40,31 @@ class TestFourthOrderTensor(unittest.TestCase):
                 for k in range(o):
                     np.testing.assert_array_equal(ab[i,j,k].matrix, a[i,j,k].matrix * b[j,k])
 
+    def test_zeros_setitem(self):
+        m, n = 4, 5
+        t = FourthOrderTensor.zeros()
+        assert t.shape == ()
+        assert np.all(t.full_tensor()==0.)
+
+        t = FourthOrderTensor.zeros(n)
+        assert t.shape == (n,)
+
+        t = FourthOrderTensor.zeros((m,n))
+        assert t.shape == (m, n)
+
+        t[1,3] = np.ones((6,6))
+        for i in range(m):
+            for j in range(n):
+                if (i == 1) and (j == 3):
+                    assert np.all(t[i,j] == 1.)
+                else:
+                    assert np.all(t[i, j] == 0.)
+
+        t0 = t == 0.
+        t0_th = np.ones((m, n))
+        t0_th[1,3] = 0.
+        assert np.all(t0== t0_th)
+
 
 class TestSymmetricFourthOrderTensor(unittest.TestCase):
     def test_inversion(self):
