### 2025-03-11 acc3392
Throw error is rhs is not suitable for setting
