### 2025-03-11 9dc076e
Fix error when trying to compute S*stress
