### 2025-03-11 32491d2
Add axis option for averages
