+++ src/Elasticipy/tensors/elasticity.py
@@ -628,10 +628,17 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
         """
         return self.inv().bulk_modulus
 
-    def Voigt_average(self):
+    def Voigt_average(self, axis=None):
         """
-        Compute the Voigt average of the stiffness tensor. If the tensor contains no orientation, we assume isotropic
-        behaviour. Otherwise, the mean is computed over all orientations.
+        Compute the Voigt average of the stiffness tensor.
+
+        If the tensor is a tensor array, all its values are considered. Otherwise (i.e. if single), the corresponding
+        isotropic tensor is returned.
+
+        Parameters
+        ----------
+        axis : int, optional
+            If provided, the average is computed along this axis. Otherwise, the mean is computed on the flattened array.
 
         Returns
         -------
@@ -645,7 +652,7 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
         average : generic function for calling either the Voigt, Reuss or Hill average
         """
         if self.ndim:
-            return self.mean()
+            return self.mean(axis=axis)
         else:
             c = self.matrix
             C11 = (c[0, 0] + c[1, 1] + c[2, 2]) / 5 \
@@ -658,10 +665,15 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
             mat = _isotropic_matrix(C11, C12, C44)
             return StiffnessTensor(mat, symmetry='isotropic', phase_name=self.phase_name)
 
-    def Reuss_average(self):
+    def Reuss_average(self, axis=None):
         """
         Compute the Reuss average of the stiffness tensor. If the tensor contains no orientation, we assume isotropic
         behaviour. Otherwise, the mean is computed over all orientations.
+
+        Parameters
+        ----------
+        axis : int, optional
+            If provided, axis to compute the average along with. If none, the average is computed on the flattened array
 
         Returns
         -------
@@ -674,12 +686,17 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
         Hill_average : compute the Voigt-Reuss-Hill average
         average : generic function for calling either the Voigt, Reuss or Hill average
         """
-        return self.inv().Reuss_average().inv()
+        return self.inv().Reuss_average(axis=axis).inv()
 
-    def Hill_average(self):
+    def Hill_average(self, axis=None):
         """
         Compute the (Voigt-Reuss-)Hill average of the stiffness tensor. If the tensor contains no orientation, we assume
         isotropic behaviour. Otherwise, the mean is computed over all orientations.
+
+        Parameters
+        ----------
+        axis : int, optional
+            If provided, axis to compute the average along with. If none, the average is computed on the flattened array
 
         Returns
         -------
@@ -692,11 +709,11 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
         Reuss_average : compute the Reuss average
         average : generic function for calling either the Voigt, Reuss or Hill average
         """
-        Reuss = self.Reuss_average()
-        Voigt = self.Voigt_average()
+        Reuss = self.Reuss_average(axis=axis)
+        Voigt = self.Voigt_average(axis=axis)
         return (Reuss + Voigt) * 0.5
 
-    def average(self, method):
+    def average(self, method, axis=None):
         """
         Compute either the Voigt, Reuss, or Hill average of the stiffness tensor.
 
@@ -704,8 +721,10 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
 
         Parameters
         ----------
+        axis : int, optional
[…]
