### 2025-03-12 d1114bd
Add test for Voigt and Reuss averages with axis argument
