### 2025-03-12 61b9313
Check that the averages are independent of the axis order
