### 2025-03-12 6ff90f6
Copy test from Stiffness to Compliance
