### 2025-03-12 3ad2b27
Avoid shadowing str
