### 2025-03-12 6df0a53
Ensure correct argument type for input matrix
