### 2025-03-12 b99c7b7
Make bulk_modulus suitable for tensor array
