### 2025-03-12 687ef8a
Also check that the bulk moduli are constant when rotating a tensor
