### 2025-03-12 0e2ca99
use _single_tensor_only() method
