+++ src/Elasticipy/tensors/elasticity.py
@@ -1099,8 +1099,7 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
         .. [3] S. I. Ranganathan and M. Ostoja-Starzewski, Universal Elastic Anisotropy Index,
            *Phys. Rev. Lett.*, 101(5), 055504, 2008. https://doi.org/10.1103/PhysRevLett.101.055504
         """
-        if self.ndim:
-            raise ValueError('The universal anisotropy factor cannot be computed on tensor arrays.')
+        self._single_tensor_only('universal_anisotropy')
         Cvoigt = self.Voigt_average()
         Creuss = self.Reuss_average()
         Gv = Cvoigt.matrix[3, 3]
