### 2025-03-12 65ce41e
Ensure that the mapping convention is kept when adding two tensors
