### 2025-03-17 19f4855
Try implementing minimization
