+++ Examples/SelfConsistent.py
@@ -1,11 +1,26 @@
 import numpy as np
-from Elasticipy.tensors.fourth_order import StiffnessTensor, ComplianceTensor, FourthOrderTensor
+from Elasticipy.tensors.fourth_order import FourthOrderTensor
+from Elasticipy.tensors.elasticity import StiffnessTensor, ComplianceTensor
 from scipy.integrate import trapezoid
 from scipy.spatial.transform import Rotation
+from scipy.integrate import dblquad
+from scipy.optimize import minimize, Bounds
 
 I = FourthOrderTensor.identity()
-#I = np.einsum('ik,jl->ijkl', np.eye(3), np.eye(3))
 global phi, theta
+
+
+def extract_upper_triangular_stiffness(C):
+    upper_triangular = C.matrix[np.triu_indices(6)]
+    return upper_triangular
+
+def reconstruct_symmetric_from_1d(upper_triangular_1d):
+    reconstructed_matrix = np.zeros((6, 6))
+    indices = np.triu_indices(6)
+    reconstructed_matrix[indices] = upper_triangular_1d
+    reconstructed_matrix = reconstructed_matrix + reconstructed_matrix.T - np.diag(reconstructed_matrix.diagonal())
+
+    return reconstructed_matrix
 
 def gamma(C_macro_local, a1=1, a2=1, a3=1):
     s1 = np.sin(theta)*np.cos(phi) / a1
@@ -22,11 +37,30 @@ def Morris_tensor(C_macro_local):
     b= trapezoid(a, theta[0], axis=0)/(4*np.pi)
     return FourthOrderTensor(b)
 
+def gamma_int(C_macro_local, theta1, phi1):
+    s1 = np.sin(theta1)*np.cos(phi1)
+    s2 = np.sin(theta1)*np.sin(phi1)
+    s3 = np.cos(theta1)
+    s = [s1, s2, s3]
+    D = np.einsum('lmnp,p,l->mn', C_macro_local.full_tensor(), s, s)
+    return np.einsum('nr,w,s->nwrs', np.linalg.inv(D), s, s)
+
+def Morris_tensor_int(C_macro_local):
+    E = np.zeros((3,3,3,3))
+    for n in range(3):
+        for w in range(3):
+            for r in range(3):
+                for s in range(3):
+                    def fun(phi1, theta1):
+                        return np.sin(theta1) * gamma_int(C_macro_local, theta1, phi1)[n,w,r,s]
+                    E[n,w,r,s] = 1/(4*np.pi) * dblquad(fun, 0, np.pi, 0, 2*np.pi)[0]
+    return FourthOrderTensor(E)
+
 def localization_tensor(C_macro_local, C_incl):
     E = Morris_tensor(C_macro_local)
-    delta = FourthOrderTensor(C_incl.full_tensor() - C_macro_local)
+    delta = FourthOrderTensor(C_incl.full_tensor() - C_macro_local.full_tensor())
     Ainv = E.ddot(delta) + I
-    return Ainv.inv()
+    return Ainv.inv().full_tensor()
 
 def global_spherical_grid(n_theta=50, n_phi=100):
     global phi, theta
@@ -34,9 +68,10 @@ def global_spherical_grid(n_theta=50, n_phi=100):
     phi = np.linspace(0, 2 * np.pi, n_phi)
     phi, theta = np.meshgrid(phi, theta, indexing='ij')
 
-def Kroner_Eshelby(C, g, method='stress', max_iter=50, atol=1e-3, rtol=1e-4, display=False):
-    C_rotated = C * g
-    C_macro = StiffnessTensor.isotropic(E=100, nu=0.3)
[…]
