### 2025-03-17 75dd468
Morris tensor has no minor symmetry
