+++ Examples/SelfConsistent.py
@@ -6,7 +6,7 @@ from scipy.spatial.transform import Rotation
 from scipy.integrate import dblquad
 from scipy.optimize import minimize, Bounds
 
-I = FourthOrderTensor.identity()
+I = FourthOrderTensor.identity(mapping='Voigt')
 global phi, theta
 
 
@@ -35,7 +35,7 @@ def Morris_tensor(C_macro_local):
     gsin = (g.T*np.sin(theta.T)).T
     a = trapezoid(gsin, phi[:,0], axis=0)
     b= trapezoid(a, theta[0], axis=0)/(4*np.pi)
-    return FourthOrderTensor(b)
+    return b
 
 def gamma_int(C_macro_local, theta1, phi1):
     s1 = np.sin(theta1)*np.cos(phi1)
@@ -51,27 +51,31 @@ def Morris_tensor_int(C_macro_local):
         for w in range(3):
             for r in range(3):
                 for s in range(3):
-                    def fun(phi1, theta1):
-                        return np.sin(theta1) * gamma_int(C_macro_local, theta1, phi1)[n,w,r,s]
+                    def fun(phi_y, theta_x):
+                        return np.sin(theta_x) * gamma_int(C_macro_local, theta_x, phi_y)[n,w,r,s]
                     E[n,w,r,s] = 1/(4*np.pi) * dblquad(fun, 0, np.pi, 0, 2*np.pi)[0]
-    return FourthOrderTensor(E)
+    return E
 
 def localization_tensor(C_macro_local, C_incl):
     E = Morris_tensor(C_macro_local)
     delta = FourthOrderTensor(C_incl.full_tensor() - C_macro_local.full_tensor())
-    Ainv = E.ddot(delta) + I
+    Ainv = FourthOrderTensor(np.einsum('ijmn,mnkl->ijkl', E, delta.full_tensor())) + I
+    Sesh= np.einsum('ijmn,mnkl->ijkl', E, C_macro_local.full_tensor())
+    nu = C_macro_local.Poisson_ratio.mean()
+    beta = 2*(4-5*nu) / (15 * (1 - nu))
+    Sesh_th_full = (1-2*beta) * np.einsum('ij,kl->ijkl', np.eye(3), np.eye(3)) + beta * I.full_tensor()
+    Sesh_th = FourthOrderTensor(Sesh_th_full, mapping='Voigt')
     return Ainv.inv().full_tensor()
 
-def global_spherical_grid(n_theta=50, n_phi=100):
+def global_spherical_grid(n_theta=100, n_phi=100):
     global phi, theta
     theta = np.linspace(0, np.pi, n_theta)
     phi = np.linspace(0, 2 * np.pi, n_phi)
     phi, theta = np.meshgrid(phi, theta, indexing='ij')
 
 def Kroner_Eshelby(Ci, g, max_iter=5, atol=1e-3, rtol=1e-4, display=False):
-    theta = 0.1
     Ci_rotated = (Ci * g)
-    C_macro = Ci_rotated.Hill_average()
+    C_macro = Ci.Hill_average()
     eigen_stiff = C_macro.eig_stiffnesses
     global_spherical_grid()
     keep_on = True
@@ -84,10 +88,10 @@ def Kroner_Eshelby(Ci, g, max_iter=5, atol=1e-3, rtol=1e-4, display=False):
         C_macro_local = C_macro * (g.inv())
         for i in range(m):
             A_local[i] = localization_tensor(C_macro_local[i], Ci)
-        CiAi_local = Ci.ddot(A_local)
-        CiAi = CiAi_local * g
+        A = A_local * g
+        CiAi = Ci_rotated.ddot(A)
[…]
