### 2025-03-17 222bf20
Simplify structure
