+++ Examples/SelfConsistent.py
@@ -1,83 +1,42 @@
 import numpy as np
 from Elasticipy.tensors.fourth_order import FourthOrderTensor
-from Elasticipy.tensors.elasticity import StiffnessTensor, ComplianceTensor
+from Elasticipy.tensors.elasticity import StiffnessTensor
 from scipy.integrate import trapezoid
 from scipy.spatial.transform import Rotation
-from scipy.integrate import dblquad
-from scipy.optimize import minimize, Bounds
 
-I = FourthOrderTensor.identity(mapping='Voigt')
-global phi, theta
+I = FourthOrderTensor.identity()
 
-
-def extract_upper_triangular_stiffness(C):
-    upper_triangular = C.matrix[np.triu_indices(6)]
-    return upper_triangular
-
-def reconstruct_symmetric_from_1d(upper_triangular_1d):
-    reconstructed_matrix = np.zeros((6, 6))
-    indices = np.triu_indices(6)
-    reconstructed_matrix[indices] = upper_triangular_1d
-    reconstructed_matrix = reconstructed_matrix + reconstructed_matrix.T - np.diag(reconstructed_matrix.diagonal())
-
-    return reconstructed_matrix
-
-def gamma(C_macro_local, a1=1, a2=1, a3=1):
+def gamma(C_macro_local, phi, theta, a1, a2, a3):
     s1 = np.sin(theta)*np.cos(phi) / a1
     s2 = np.sin(theta)*np.sin(phi) / a2
     s3 = np.cos(theta) / a3
-    s = [s1, s2, s3]
+    s = np.array([s1, s2, s3])
     D = np.einsum('lmnp,pqr,lqr->qrmn', C_macro_local.full_tensor(), s, s)
     return np.einsum('qrik,jqr,lqr->qrijkl', np.linalg.inv(D), s, s)
 
-def Morris_tensor(C_macro_local):
-    g = gamma(C_macro_local)
+def polarization_tensor(C_macro_local, phi, theta, a1=1, a2=1, a3=1):
+    g = gamma(C_macro_local, phi, theta, a1, a2, a3)
     gsin = (g.T*np.sin(theta.T)).T
     a = trapezoid(gsin, phi[:,0], axis=0)
     b= trapezoid(a, theta[0], axis=0)/(4*np.pi)
     return b
 
-def gamma_int(C_macro_local, theta1, phi1):
-    s1 = np.sin(theta1)*np.cos(phi1)
-    s2 = np.sin(theta1)*np.sin(phi1)
-    s3 = np.cos(theta1)
-    s = [s1, s2, s3]
-    D = np.einsum('lmnp,p,l->mn', C_macro_local.full_tensor(), s, s)
-    return np.einsum('nr,w,s->nwrs', np.linalg.inv(D), s, s)
-
-def Morris_tensor_int(C_macro_local):
-    E = np.zeros((3,3,3,3))
-    for n in range(3):
-        for w in range(3):
-            for r in range(3):
-                for s in range(3):
-                    def fun(phi_y, theta_x):
-                        return np.sin(theta_x) * gamma_int(C_macro_local, theta_x, phi_y)[n,w,r,s]
-                    E[n,w,r,s] = 1/(4*np.pi) * dblquad(fun, 0, np.pi, 0, 2*np.pi)[0]
-    return E
-
-def localization_tensor(C_macro_local, C_incl):
-    E = Morris_tensor(C_macro_local)
-    delta = FourthOrderTensor(C_incl.full_tensor() - C_macro_local.full_tensor())
[…]
