### 2025-03-17 aa36cc6
The aspect ratio are passed from localization_tensor
