### 2025-03-17 6e90d8f
Use from_Kelvin instead of full tensor
