### 2025-03-18 5ed2d6b
Fix confusion between phi and phi_grid
