### 2025-03-18 a08fedf
Implement polarization definition from Masson (2008)
