### 2025-03-19 e88f401
Add option to apply minor symmetry on-the-fly
