### 2025-03-19 f8db0bc
Use matrices instead of full_tensor
