+++ Examples/SelfConsistent.py
@@ -53,9 +53,8 @@ def polarization_tensor(C_macro_local, a1, a2, a3, n_phi, n_theta):
 
 def localization_tensor(C_macro_local, C_incl, n_phi, n_theta):
     E = polarization_tensor(C_macro_local, 0.1, 1, 10, n_phi, n_theta)
-  #  E2 = Morris_tensor_int(C_macro_local, 10, 1, 0.1)
-    delta = C_incl.full_tensor() - C_macro_local.full_tensor()
-    Ainv = FourthOrderTensor(np.einsum('ijmn,mnkl->ijkl', E, delta)) + I
+    delta = FourthOrderTensor(C_incl.matrix - C_macro_local.matrix)
+    Ainv = E.ddot(delta) + I
     return Ainv.inv().full_tensor()
 
 def Kroner_Eshelby(Ci, g, max_iter=5, atol=1e-3, rtol=1e-3, display=False, n_phi=100, n_theta=100):
