### 2025-03-19 f0de2fc
Symmetrize the tensor afterward
