### 2025-03-19 ca8a2a5
Remove unused functions
