### 2025-03-19 2b7dfdd
Set index value form matrix, not from full_tensor
