### 2025-03-19 4cb9a94
Add default checking for minor symmetry
