### 2025-03-19 f76f1e8
Add test for symmetry checking
