### 2025-03-19 91011b9
Rename option 'check_symmetry' > check_symmetries and 'force_symmetry' > 'force_symmetries'
