### 2025-03-20 5de8228
Implement invariants
