### 2025-03-20 a3a5ea5
Properly format math
