### 2025-03-21 09abc55
Implement is_isotropic, is_cubic and is_tetragonal
