### 2025-03-21 334142c
Test is_isotropic()
