### 2025-03-21 dbcf1ba
Add tests for is_cubic and is_tetragonal
