### 2025-03-21 9d463f2
Add tests for invariants
