### 2025-03-21 32f0589
Implement negative unary operator
