### 2025-03-21 ade28e6
Fix atol option (not passed to inner function)
