### 2025-03-21 111ad7f
Use symmetrization from tensor constructor
