### 2025-03-22 #7940 @editorialbot
[REVIEW]: Elasticipy: A Python package for linear elasticity and tensor analysis
**Submitting author:** @DorianDepriester (<a href="http://orcid.org/0000-0002-2881-8942">Dorian Depriester</a>)
**Repository:** https://github.com/DorianDepriester/Elasticipy
**Branch with paper.md** (empty if default branch):
**Version:** v4.2.0
