### 2025-03-25 afe8562
eig_signature also returns unique eigen values
