### 2025-03-25 b93cab4
Fix wrong equation formating
