### 2025-03-25 e0583af
Zener ratio is now independent of the orientation
