### 2025-03-25 d0e59a6
Add test for Zener ratio for rotated tensor
