### 2025-03-28 8aa8723
Rename _eig_signature to public method eig_stiffnesses_multiplicity
