### 2025-03-28 534aa1d
Mention eig_stiffnesses_multiplicity in See Also section
