### 2025-03-28 835911c
Add docstring and doctest
