### 2025-03-28 6dc55bd
use orders, as the unique eigenstiffness are no longer sorted by increasing multiplicity
