### 2025-03-28 344b0ce
Add tol argument for Zener_ratio
