### 2025-03-28 c94618d
Zener_ratio is no longer a property Check that an error is thrown if the material has no cubic symmetry
