### 2025-04-01 d441d7a
is_cubic, is_isotropic and is_tetragonal also work for tensor arrays
