### 2025-04-01 4cc4ac7
Move test_isotropic to TestStiffnessConstructor
