### 2025-04-01 fdb3fd7
Test is_cubic for nD tensor array
