### 2025-04-01 6df1748
isotropy => cubic symmetry
