### 2025-04-01 167f8f4
Also check that an isotropic stiffness tensor has cubic symmetry (and so for tetragonal and cubic)
