### 2025-04-01 ba408d3
Fix error when converting array to set (which removes duplicates)
