### 2025-04-01 e892294
Fix error when converting array to set (which removes duplicates) for arrays
