### 2025-04-01 5bf7c65
Add test for is_isotropic, is_cubic and is_tetragonal for rotated tensors
