### 2025-04-04 0c0ddca
Fix typo in docstring
