### 2025-04-10 f99656f
Add test for deprecation warnings
