+++ tests/test_StiffnessTensor.py
@@ -283,6 +283,14 @@ class TestComplianceTensor(unittest.TestCase):
         for i, inv_rotated_i in enumerate(S_rotated.invariants()):
             np.testing.assert_array_almost_equal(inv_rotated_i, inv[i])
 
+    def test_deprecated_path(self):
+        expected_warn = ("The module 'Elasticipy.FourthOrderTensor' is deprecated and will be removed in a future "
+                            "release. Please use 'Elasticipy.tensors.elasticity' instead.")
+        with self.assertWarns(DeprecationWarning) as context:
+            from Elasticipy.FourthOrderTensor import ComplianceTensor
+        self.assertEqual(str(context.warning), expected_warn)
+
+
 class TestStiffnessConstructor(unittest.TestCase):
     def test_averages(self):
         """Check that the Voigt, Reuss and Hill averages are consistent with those provided by MP."""
@@ -940,6 +948,13 @@ class TestStiffnessConstructor(unittest.TestCase):
         C_rotated = C * rotations
         for i, inv_rotated_i in enumerate(C_rotated.invariants()):
             np.testing.assert_array_almost_equal(inv_rotated_i, inv[i])
+
+    def test_deprecated_path(self):
+        expected_warn = ("The module 'Elasticipy.FourthOrderTensor' is deprecated and will be removed in a future "
+                            "release. Please use 'Elasticipy.tensors.elasticity' instead.")
+        with self.assertWarns(DeprecationWarning) as context:
+            from Elasticipy.FourthOrderTensor import StiffnessTensor
+        self.assertEqual(str(context.warning), expected_warn)
 
 if __name__ == '__main__':
     unittest.main()
