### 2025-04-10 700fa1f
Add test for deprecation warning
