### 2025-04-10 922fb8e
Add test for deprecation warning
