### 2025-04-10 c8bfc05
No deprecation warning is raised if it was raised before
