### 2025-04-10 f2da29e
Fix extra tabs in doctest
