### 2025-04-10 e25aee2
Update README.md
Add citation section
