### 2025-05-09 8ee7173
Allow to define Poisson ratios in two ways (nu_xy or nu_yx)
