### 2025-05-09 023204a
Test orthotropic with nu_xy, nu_yx, or both
