### 2025-05-09 c834c4a
Allow to define Poisson ratios in two ways (nu_xy or nu_yx)
