### 2025-05-09 7a3313d
Test transverse isotropic with two ways for the Poisson ratios (eg. nu_xy and nu_yx)
