### 2025-05-09 ea626d7
Double check that the definition for the Poisson ratios are consistent with naming convention
