### 2025-05-09 7c2fcd3
Show matrix convention in notes
