### 2025-07-03 d88d628
Use qtpy instead of PyQt (fix #4)
