+++ requirements.txt
@@ -3,3 +3,4 @@ scipy
 matplotlib
 setuptools
 pandas
+qtpy
