### 2025-07-03 649813f
Add pandas to dependencies (fix #5)
