+++ .readthedocs.yaml
@@ -15,7 +15,7 @@ build:
     # golang: "1.20"
   jobs:
     post_create_environment:
-      - python -m pip install sphinx_rtd_theme mp_api
+      - python -m pip install sphinx_rtd_theme mp_api sphinx_copybutton
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
