+++ src/Elasticipy/tensors/elasticity.py
@@ -89,6 +89,13 @@ class StiffnessTensor(SymmetricFourthOrderTensor):
             Whether to check or not that the input matrix is symmetric.
         force_symmetry : bool, optional
             If true, the major symmetry of the tensor is forces
+
+        Notes
+        -----
+        The units used when building the stiffness tensor are up to the user (GPa, MPa, dyn etc.). Therefor, the
+        results you will get when performing operations (Young's modulus, "product" with strain tensor etc.) will be
+        consistent with the these units. For instance, if the stiffness tensor is defined in GPa, the computed stress
+        will be given in GPa as well.
         """
         super().__init__(M, mapping=mapping, **kwargs)
         if check_positive_definite:
