### 2025-07-08 e2d30d0
Add note about the units (fix #8)
