### 2025-07-08 19224fd
Add note about the units (fix #8)
