### 2025-07-08 24d6a73
Try to fix issue with prompt symbol >>> and ...
