+++ CONTRIBUTING.md
@@ -25,7 +25,7 @@ All types of contributions are encouraged and valued. See the [Table of Contents
 ## Code of Conduct
 
 This project and everyone participating in it is governed by the
-[Elasticipy Code of Conduct](https://github.com/DorianDepriester/Elasticipy/blob//CODE_OF_CONDUCT.md).
+[Elasticipy Code of Conduct](https://github.com/DorianDepriester/Elasticipy/blob/main/CODE_OF_CONDUCT.md).
 By participating, you are expected to uphold this code. Please report unacceptable behavior
 to <dorian.dep@gmail.com>.
 
