### 2025-07-08 c88d66a
Remove plt.show() [related to #12]
