### 2025-07-08 8d694fa
Add plt.show(), as it has been removed from plotting functions
