### 2025-07-08 0c27acd
Add missing fig.show()
