### 2025-07-09 b7fce5c
Remove unused show argument
