+++ docs/source/Tutorials/Tutorial_wave-velocities.rst
@@ -46,7 +46,7 @@ Now, we can define 3 spherical functions, which correspond to velocities of:
 
     >>> cp, cs_fast, cs_slow = C.wave_velocity(rho)
 
-One can check that for a given direction (say, **x**), the velocities are arranged in the right order:
+One can check that for a given direction (say, **x**), the velocities are sorted in the right order:
 
     >>> x = [1, 0, 0]
     >>> print(cp.eval(x), cs_fast.eval(x), cs_slow.eval(x))
