### 2025-07-09 7f7d158
Remove show option, and add fig.show()
