### 2025-07-09 4adcc94
Drop (obsolete) argument 'show'
