### 2025-07-09 69b68a0
Remove obsolete 'show' argument
