### 2025-07-09 40592f8
Add string representations
