### 2025-07-09 014d7a7
Add test for __repr__
