### 2025-07-09 9b0fce6
Fix wrong name field for class
