### 2025-07-09 075a7ba
fix BP 2 in https://github.com/openjournals/joss-reviews/issues/7940#issuecomment-3027326768
