### 2025-07-09 f0a9419
fix BP 3 in https://github.com/openjournals/joss-reviews/issues/7940#issuecomment-3027326768
