+++ JOSS/paper.md
@@ -136,6 +136,7 @@ to other averaging methods (e.g. self-consistent models), possibly beyond linear
 with ease. It already implements thermal expansion.
 
 # Usage
+This section presents the syntaxes of few basic operations performed with Elasticipy v4.0.0.
 
 ## Plot directional engineering constants
 
