### 2025-07-09 92969c2
fix BP 4 in https://github.com/openjournals/joss-reviews/issues/7940#issuecomment-3027326768
