### 2025-07-09 8c801d9
Remove import matplotlib, use fig.show() instead
