### 2025-07-09 0ca53fd
Add space around = setter
