### 2025-07-09 4c5aeb3
Remove blank lines in all code snippets
