### 2025-07-09 45e3a5c
Try adding horizontal lines around code blocks
