### 2025-07-09 a94fc2a
Revert "Try adding horizontal lines around code blocks"
This reverts commit 45e3a5ce9a0f644a2915032632332ab360780b79.
