### 2025-07-10 6928696
Fix false negative doctest
