### 2025-07-10 8ac2fe3
Fix false negative doctest
