### 2025-07-10 9e9d432
The output arguments are arrays
