+++ src/Elasticipy/interfaces/prisms_plasticity.py
@@ -6,7 +6,7 @@ import numpy as np
 def rebuild_tensor(F11, F22, F33, F12, F13, F21, F23, F31, F32):
     return np.array([[F11, F12, F13], [F21, F22, F23], [F31, F32, F33]]).transpose((2,0,1))
 
-def from_quadrature(file, returns='stress'):
+def from_quadrature_file(file, returns='stress'):
     """
     Read data from quadrature output file generated by PRISMS plasticity.
 
