### 2025-07-10 98d9902
Make rebuild_tensor private
