### 2025-07-10 f0ed6fe
Add from_stresstrain_file function
