+++ src/Elasticipy/interfaces/prisms_plasticity.py
@@ -1,4 +1,4 @@
-from Elasticipy.tensors.second_order import SecondOrderTensor
+from Elasticipy.tensors.second_order import SecondOrderTensor, SymmetricSecondOrderTensor
 from Elasticipy.tensors.stress_strain import StressTensor
 import pandas as pd
 import numpy as np
@@ -72,4 +72,32 @@ def from_quadrature_file(file, returns='stress'):
     else:
         return tuple(returned_values)
 
+def from_stressstrain_file(file):
+    """
+    Read data from stress-strain file generated by PRISMS plasticity.
+
+    This file is usually named stressstrain.txt
+
+    Parameters
+    ----------
+    file : str
+        Path to stress-strain file
+
+    Returns
+    -------
+    SymmetricSecondOrderTensor
+        Average Green-Lagrange strain tensor
+    StressTensor
+        Average stress tensor
+    """
+    data = pd.read_csv(file, sep='\t')
+    Exx, Exy, Exz = data['Exx'], data['Exy'], data['Exz']
+    Eyy, Eyz = data['Eyy'], data['Eyz']
+    Ezz = data['Ezz']
+    E = np.array([[Exx, Exy, Exz], [Exy, Eyy, Eyz], [Exz, Eyz, Ezz]]).transpose((2,0,1))
+    Txx, Txy, Txz = data['Txx'], data['Txy'], data['Txz']
+    Tyy, Tyz = data['Tyy'], data['Tyz']
+    Tzz = data['Tzz']
+    T = np.array([[Txx, Txy, Txz], [Txy, Tyy, Tyz], [Txz, Tyz, Tzz]]).transpose((2, 0, 1))
+    return SymmetricSecondOrderTensor(E), StressTensor(T)
 
