### 2025-07-10 37139b3
Add test_interfaces_prisms.py
