### 2025-07-10 4c8b8e6
Fix wrong file extension in docstring long description
