+++ src/Elasticipy/interfaces/prisms_plasticity.py
@@ -10,7 +10,7 @@ def from_quadrature_file(file, returns='stress'):
     """
     Read data from quadrature output file generated by PRISMS plasticity.
 
-    These files, usually named QuadratureOutputsXXX.txt (where XXX denotes the time increment), contain large amount of
+    These files, usually named QuadratureOutputsXXX.csv (where XXX denotes the time increment), contain large amount of
     data, such as gradient components, stress values, grain ID etc.
 
     Parameters
