+++ tests/test_interfaces_prisms.py
@@ -8,14 +8,19 @@ import os
 current_dir = os.path.dirname(os.path.abspath(__file__))
 quadrature_file = os.path.join(current_dir, 'QuadratureOutputs.csv')
 
-class TestPRISMInterfaces(unittest.TestCase):
+quadrature_data = pd.read_csv(quadrature_file, header=None, usecols=range(0,37))
+
 class TestPRISMSInterfaces(unittest.TestCase):
     def test_from_quadrature(self):
         stress = from_quadrature_file(quadrature_file)
         assert isinstance(stress, StressTensor)
-        assert stress[0].C[0, 0] == 1.59893e+01
-        assert stress[0].C[1, 1] == 7.76097
-        assert stress[0].C[2, 2] == 6.51261
+        for i in range(0,len(quadrature_data)):
+            assert stress[i].C[0, 0] == quadrature_data.iloc[i,28]
+            assert stress[i].C[1, 1] == quadrature_data.iloc[i,29]
+            assert stress[i].C[2, 2] == quadrature_data.iloc[i,30]
+            assert stress[i].C[0, 1] == quadrature_data.iloc[i,31]
+            assert stress[i].C[0, 2] == quadrature_data.iloc[i,32]
+            assert stress[i].C[1, 2] == quadrature_data.iloc[i,34]
 
     def test_from_stressstrain_with_fields(self):
         fields = ('grain ID', 'phase ID', 'det(J)', 'twin', 'coordinates', 'orientation', 'elastic gradient',
