### 2025-07-10 157caee
Check size of returned data
