### 2025-07-10 b4998f0
Add test for from_stressstrain_file
