### 2025-07-11 d9b3dbc
Add stack method for SecondOrderTensor
