### 2025-07-11 fa54a5a
Add test for stack method
