### 2025-07-11 70e6bda
Add case where the given axis is negative
