### 2025-07-11 eb69f07
Also test with negative axis
