### 2025-07-11 9689516
More simple case
