### 2025-07-11 4e952a8
Add option to force Voigt mapping factors
