### 2025-07-11 091af80
Add size check
