### 2025-07-11 ae39e21
Support for skew-symmetric tensor
