### 2025-07-11 89b5b74
Add test for skew-symmetric
