### 2025-07-13 d0e55c8
Add argument to force type of return
