+++ src/Elasticipy/interfaces/FEPX.py
@@ -13,7 +13,7 @@ def _list_valid_filenames(folder, startswith='strain'):
     pattern = r'{}\.step\d+'.format(startswith)
     return [f for f in file_list if re.fullmatch(pattern, f)]
 
-def from_step_file(file):
+def from_step_file(file, dtype=None):
     """
     Import data from a single step file given by FEPX.
 
@@ -24,6 +24,16 @@ def from_step_file(file):
     ----------
     file : str
         Path to the file to read
+    dtype : str, optional
+        If provided, force sets the type of returned array. It can be:
+          - None (let the function infer the dtype from the data)
+          - SecondOrderTensor
+          - SymmetricSecondOrderTensor
+          - SkewSymmetricSecondOrderTensor
+          - stressTensor
+          - strainTensor
+          - float
+          - int
 
     Returns
     -------
@@ -34,18 +44,26 @@ def from_step_file(file):
     data = pd.read_csv(file, header=None, sep=' ')
     array = data.to_numpy()
     base_name = os.path.splitext(os.path.basename(file))[0]
-    if base_name == 'strain':
+    if isinstance(dtype, str):
+        dtype = re.sub('[^A-Za-z0-9]+', '', dtype).lower()
+    if base_name == 'strain' or dtype=='straintensor':
         return StrainTensor.from_Voigt(array, voigt_map=[1,1,1,1,1,1])
-    elif base_name == 'stress':
+    elif base_name == 'stress' or dtype=='stresstensor':
         return StressTensor.from_Voigt(array)
     else:
         n_compo = array.shape[1]
         if n_compo == 3:
-            zeros = np.zeros(array.shape[0])
-            mat = np.array([[ zeros,         array[:, 0],   array[:, 1]],
-                            [-array[:, 0],  zeros,          array[:, 2]],
-                            [-array[:, 1], -array[:, 2],    zeros     ]]).transpose((2, 0, 1))
-            return SkewSymmetricSecondOrderTensor(mat)
+            if dtype is None:
+                raise ValueError('I cannot automatically infer the dtype from the data. Use dtype option to set whether'
+                                 ' it is a float array or a skew-symmetric second-order tensor array.')
+            elif dtype == 'skewsymmetricsecondordertensor':
+                zeros = np.zeros(array.shape[0])
+                mat = np.array([[ zeros,         array[:, 0],   array[:, 1]],
+                                [-array[:, 0],  zeros,          array[:, 2]],
+                                [-array[:, 1], -array[:, 2],    zeros     ]]).transpose((2, 0, 1))
+                return SkewSymmetricSecondOrderTensor(mat)
+            elif dtype == 'float':
+                return array
         elif n_compo == 6:
             return SymmetricSecondOrderTensor.from_Voigt(array)
         elif n_compo == 9:
