### 2025-07-13 6b278fb
Add argument to force type of return
