### 2025-07-13 07dda32
Fix wrong double call for from_results_file
