### 2025-07-13 c7c7603
Add test for ambiguous dtype
