### 2025-07-15 4189e0a
Define a full table for fetching the proper constructor
