### 2025-07-15 724d238
Remove error testing (no longer thrown)
