### 2025-07-15 51b70de
Add test for numpy array (with ori)
