### 2025-07-17 236663d
Implement draw_Mohr_circles
