### 2025-07-17 e9be864
Add docstring
