### 2025-07-17 1411271
Remove duplicates in xticks and yticks, if any
