### 2025-07-17 a1cfcbf
Add test for draw_Mohr_circles()
