### 2025-07-17 c102722
Add section for the Mohr circles
