### 2025-07-17 4f199ce
Add missing return type
