### 2025-07-17 3cfc1a0
Implement mode='cross'
