### 2025-07-17 efcfed8
Add docstring for add_polefigure
