### 2025-07-17 60a09ab
Add analysis section after the Mohr circles
