### 2025-07-20 506d8dc
Fix wrong path to div API doc (used to point to former path)
