### 2025-09-01 c62ab73
More straightfoward way to compute the shear modulus
