### 2025-09-01 a0c60db
More straightfoward way to compute the Young modulus
