### 2025-09-01 515b350
More straightfoward way to compute the Poisson ratio
