### 2025-09-01 ea627cd
More straightfoward way to compute the linear compressibility
