### 2025-09-01 8f714a8
Remove obsolete function
