### 2025-11-03 #7940 @AHartmaier
@DorianDepriester I still have an issue with PyQt5 after installation of Elasticipy with pip. All other issues are resolved now.
