### 2025-11-05 #7940 @AHartmaier
@DorianDepriester Do you plan to publish the package also on conda-forge? I could trigger/support this process.
