### 2025-11-13 #7940 @DorianDepriester
Dear @srmnitc ,

- I have double-checked the authors and affiliations
- I have published a new release of Elasticipy, which includes the present reviews. It is referred to as version **4.2.0**.
- This version has been archived on [Zenodo](https://zenodo.org/records/17598307). The associated DOI is **[10.5281/zenodo.17598307](https://doi.org/10.5281/zenodo.17598307)**
- The author list and their ORCID are the same as in the JOSS paper
- The license listed on Zenodo is the same as on the repo (MIT License)
