### 2025-11-25 #7940 @Kevin-Mattheus-Moerman
-> No, needs to be changed to `v4.2.0`

- [ ] The paper mixes American English (e.g. `visualize`, `generalized`, `analyze`, `visualization`) and also British English (e.g. `analysing`), please use one form consistently.
