### 2025-11-25 #7940 @DorianDepriester
@Kevin-Mattheus-Moerman,
The paper now fully complies with UK english.
